requirements.txt
azure_storage==0.37.0
langchain_core==0.3.75
langchain_openai==0.3.32
langgraph==0.6.6
pandas==2.3.2
pyspark==4.0.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
